brokedown the navbar into multiple components, main nav file was getting to clustered this will help clean it up
