+++ src/components/ErrorPage.jsx
@@ -8,7 +8,9 @@ function ErrorPage() {
     <Container className="error-page text-center d-flex flex-column justify-content-center align-items-center">
       <h1 className="display-4">Oops!</h1>
       <p className="lead">The page you're looking for doesn't exist.</p>
-      <Button variant="primary" onClick={() => navigate('/')}>Return to Home</Button>
+      <Button variant="primary" onClick={() => navigate('/')}>
+        Return to Home
+      </Button>
     </Container>
   );
 }
