+++ src/components/Navbar/Navbar.jsx
@@ -1,30 +1,12 @@
 import Container from 'react-bootstrap/Container';
-import Nav from 'react-bootstrap/Nav';
 import Navbar from 'react-bootstrap/Navbar';
-import NavDropdown from 'react-bootstrap/NavDropdown';
-import Image from 'react-bootstrap/Image';
-import Button from 'react-bootstrap/Button';
-import OverlayTrigger from 'react-bootstrap/OverlayTrigger';
-import Tooltip from 'react-bootstrap/Tooltip';
 
-import { FaHeart } from 'react-icons/fa';
+import UserIcon from './UserIcon';
+import NavMenu from './NavMenu';
 
-import ProfilePlaceholder from '../../assets/Images/profilePlaceholder.jpg';
-
-import { useUser } from '../../Data/UserContext';
 import './Navbar.css';
 
 function CustomNavbar() {
-  const { isAuthenticated, logout } = useUser();
-
-  // Placeholder data — eventually pulled from real user object
-  const user = {
-    name: 'username',
-    email: 'username@example.com',
-    profileImg: ProfilePlaceholder,
-    role: 'Admin',
-  };
-
   return (
     <Navbar expand="lg" className="bg-body-tertiary px-3">
       <Container fluid>
@@ -33,82 +15,11 @@ function CustomNavbar() {
           <Navbar.Brand href="/">Alumbiz</Navbar.Brand>
 
           {/* Center-aligned menu */}
-          <Nav className="nav-center">
-            <Nav.Link href="/link">Why Donate Stock</Nav.Link>
-            <Nav.Link href="/link">For Donors</Nav.Link>
-            <Nav.Link href="/link">For Universities</Nav.Link>
-            <Nav.Link href="/link">Partners</Nav.Link>
-            <Nav.Link href="/link">Resources</Nav.Link>
-            <Nav.Link href="/link">Partners</Nav.Link>
-
-            {!isAuthenticated && (
-              <>
-                <Nav.Link href="/signup">Sign In</Nav.Link>
-                <OverlayTrigger
-                  placement="bottom"
-                  overlay={<Tooltip id="donate-tooltip">Donate</Tooltip>}
-                >
-                  <Button
-                    href="/donate"
-                    variant="primary"
-                    className="donate-button"
-                  >
-                    <FaHeart className="donate-heart-icon" />
-                    Donate Shares
-                  </Button>
-                </OverlayTrigger>
-              </>
-            )}
-          </Nav>
+          <NavMenu />
 
           {/* Right-aligned user profile */}
           <div className="d-flex align-items-center">
-            <Navbar.Toggle aria-controls="basic-navbar-nav" />
-            <Navbar.Collapse id="basic-navbar-nav">
-              <Nav>
-                {isAuthenticated && (
-                  <div className="d-flex align-items-center gap-2">
-                    <div className="d-flex flex-column text-end">
-                      <span className="user-email">{user.email}</span>
-                      <span className="user-role">{user.role}</span>
-                    </div>
-                    <NavDropdown
-                      align="end"
-                      title={
-                        <Image
-                          src={user.profileImg}
-                          roundedCircle
-                          width={40}
-                          height={40}
-                          alt="Profile"
-                          className="profile-icon"
-                        />
-                      }
-                      id="profile-dropdown"
-                      className="profile-dropdown"
-                    >
-                      <div className="profile-card">
-                        <Image
-                          src={user.profileImg}
-                          className="profile-card-img"
-                          alt="Profile"
-                        />
-                        <div className="profile-card-info">
-                          <strong>{user.name}</strong>
-                          <div className="text-muted">{user.email}</div>
-                        </div>
-                      </div>
-                      <NavDropdown.Item href="/profile">
-                        Profile
-                      </NavDropdown.Item>
-                      <NavDropdown.Item onClick={logout}>
-                        Sign Out
-                      </NavDropdown.Item>
-                    </NavDropdown>
-                  </div>
-                )}
-              </Nav>
-            </Navbar.Collapse>
+            <UserIcon />
           </div>
         </div>
       </Container>
